+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 * Fixed `cargo init` giving unexpected results by upgrading to `rbx_dom_weak` 1.1.0
+* Updated default place file:
+    * Improved default properties to be closer to Studio's built-in 'Baseplate' template
+    * Added a baseplate to the project file (Thanks, [@AmaranthineCodices](https://github.com/AmaranthineCodices/)!)
 
 ## [0.5.0 Alpha 5](https://github.com/LPGhatguy/rojo/releases/tag/v0.5.0-alpha.5) (March 1, 2019)
 * Upgraded core dependencies, which improves compatibility for lots of instance types